Added LineAlert component -moved alerts to own directory -renamed login in RestClient
