+++ src/rest/RestClient.ts
@@ -12,9 +12,19 @@ export default class RestClient {
       },
     });
   }
-  login = async (payload: LoginDataInterface) => {
-    //TODO should return iterface or type describing the response data, also handles the error
-    return await this.#axiosInstance.post(endpoint.loginEndpoint, JSON.stringify(payload));
+  loginRequest = (payload: LoginDataInterface) => {
+    const actionPost = async () => {
+      return await this.#axiosInstance.post(endpoint.loginEndpoint, JSON.stringify(payload));
+    };
+
+    //TODO should return interface or type describing the response data, also handles the error
+    actionPost()
+      .then((response: any) => {
+        return response;
+      })
+      .catch((err: any) => {
+        return err;
+      });
   };
 
   #axiosInstance: AxiosInstance;
